refactor: GameState abstract class 추가
